Correct isTVOS implementation
- Refactor resolveScheme, resolveArtifactPath, resolveBuildConfiguration
- Use methods above correctly in isTVOS()
